engine: fix double-free in program cache

+++ filament/src/MaterialDefinition.cpp
@@ -133,7 +133,7 @@ void releaseProgramsImpl(FEngine& engine, utils::Slice<Handle<HwProgram>> progra
                 globalProgramCache.release(specialization, [&engine](Handle<HwProgram> p) {
                     engine.getDriverApi().destroyProgram(p);
                 });
-            } else {
+            } else if (program) {
                 engine.getDriverApi().destroyProgram(program);
             }
             program.clear();
@@ -152,7 +152,7 @@ void releaseProgramsImpl(FEngine& engine, utils::Slice<Handle<HwProgram>> progra
                 globalProgramCache.release(specialization, [&engine](Handle<HwProgram> p) {
                     engine.getDriverApi().destroyProgram(p);
                 });
-            } else if (destroySharedVariants) {
+            } else if (destroySharedVariants && program) {
                 engine.getDriverApi().destroyProgram(program);
             }
             program.clear();
@@ -806,7 +806,7 @@ Handle<HwProgram> MaterialDefinition::prepareProgram(FEngine& engine, DriverApi&
     if (!hasVariant(specialization.variant, engine.getShaderModel(), driver.isStereoSupported())) {
         return {};
     }
-    if (engine.features.backend.enable_program_cache && parser == *mMaterialParser) {
+    if (UTILS_LIKELY(engine.features.backend.enable_program_cache && parser == *mMaterialParser)) {
         Handle<HwProgram>* program = engine.getMaterialCache().getProgramCache().get(specialization,
                 [this, &engine, &parser, &specialization, priorityQueue]() {
                     return compileProgram(engine, parser, specialization, priorityQueue);
@@ -823,7 +823,7 @@ void MaterialDefinition::acquirePrograms(FEngine& engine,
         MaterialParser const& parser,
         utils::Slice<const backend::Program::SpecializationConstant> specializationConstants,
         bool isDefaultMaterial) const {
-    if (engine.features.backend.enable_program_cache && parser == *mMaterialParser) {
+    if (UTILS_LIKELY(engine.features.backend.enable_program_cache && parser == *mMaterialParser)) {
         acquireProgramsImpl<true>(engine, programCache, *this, parser, specializationConstants,
                 isDefaultMaterial);
     } else {
@@ -836,7 +836,7 @@ void MaterialDefinition::releasePrograms(FEngine& engine,
         utils::Slice<Handle<HwProgram>> programCache, MaterialParser const& parser,
         utils::Slice<const backend::Program::SpecializationConstant> specializationConstants,
         bool isDefaultMaterial) const {
-    if (engine.features.backend.enable_program_cache && parser == *mMaterialParser) {
+    if (UTILS_LIKELY(engine.features.backend.enable_program_cache && parser == *mMaterialParser)) {
         releaseProgramsImpl<true>(engine, programCache, *this, specializationConstants,
                 isDefaultMaterial);
     } else {
